Rename Zuora credit adjustment module for clarity
Renamed applyCreditToAccountBalance.ts to creditBalanceAdjustment.ts in the Zuora module for improved clarity and consistency. Updated all relevant imports in handler and test files to reflect the new module name.
